+++ realm/realm-library/src/androidTestObjectServer/kotlin/io/realm/mongodb/sync/SessionTests.kt
@@ -21,11 +21,13 @@ import androidx.test.platform.app.InstrumentationRegistry
 import io.realm.*
 import io.realm.TestHelper.TestLogger
 import io.realm.entities.DefaultSyncSchema
+import io.realm.entities.SyncColor
 import io.realm.entities.SyncStringOnly
 import io.realm.entities.SyncStringOnlyModule
 import io.realm.exceptions.RealmFileException
 import io.realm.exceptions.RealmMigrationNeededException
 import io.realm.kotlin.syncSession
+import io.realm.kotlin.where
 import io.realm.log.LogLevel
 import io.realm.log.RealmLog
 import io.realm.mongodb.*
@@ -34,13 +36,13 @@ import io.realm.util.ResourceContainer
 import io.realm.util.assertFailsWithMessage
 import org.bson.types.ObjectId
 import org.hamcrest.CoreMatchers
-import org.junit.After
-import org.junit.Before
-import org.junit.Rule
+import org.junit.*
+import org.junit.Ignore
 import org.junit.Test
 import org.junit.runner.RunWith
 import java.util.concurrent.TimeUnit
 import java.util.concurrent.atomic.AtomicBoolean
+import java.util.concurrent.atomic.AtomicInteger
 import kotlin.test.*
 
 @RunWith(AndroidJUnit4::class)
@@ -123,9 +125,47 @@ class SessionTests {
         }
     }
 
+    // Check that a if Seamless loss Client Reset fails the error is correctly reported.
+    @Test
+    fun errorHandler_discardUnsyncedChangesStrategy_resetErrorHandled() = looperThread.runBlocking {
+        val email = TestHelper.getRandomEmail()
+        val user: User = app.registerUserAndLogin(email, "123456")
+
+        val config = configFactory.createSyncConfigurationBuilder(user)
+            .testSchema(SyncStringOnly::class.java)
+                .syncClientResetStrategy(object: DiscardUnsyncedChangesStrategy {
+                    override fun onBeforeReset(realm: Realm) {
+                        fail("This test case was not supposed to trigger DiscardUnsyncedChangesStrategy::onBeforeReset()")
+                    }
+
+                    override fun onAfterReset(before: Realm, after: Realm) {
+                        fail("This test case was not supposed to trigger DiscardUnsyncedChangesStrategy::onAfterReset()")
+                    }
+
+                    override fun onError(session: SyncSession, error: ClientResetRequiredError) {
+                        val filePathFromError = error.originalFile.absolutePath
+                        val filePathFromConfig = session.configuration.path
+                        assertEquals(filePathFromError, filePathFromConfig)
+                        assertFalse(error.backupFile.exists())
+                        assertTrue(error.originalFile.exists())
+                        // Note, this error message is just the one created by ObjectStore for testing
+                        // The server will send a different message. This just ensures that we don't
+                        // accidentially modify or remove the message.
+                        assertEquals("Simulate Client Reset", error.message)
+                        looperThread.testComplete()
+                    }
+                })
+            .build()
+
+        val realm = Realm.getInstance(config)
+        looperThread.closeAfterTest(realm)
+        user.app.sync.simulateClientReset(realm.syncSession, ErrorCode.AUTO_CLIENT_RESET_FAILURE)
+    }
+
     // Check that a Client Reset is correctly reported.
     @Test
-    fun errorHandler_clientResetReported() = looperThread.runBlocking {
+    @Deprecated("clientResetHandler deprecated in favor of syncClientResetStrategy")
+    fun errorHandler_manualClientResetReported() = looperThread.runBlocking {
         val config = configFactory.createSyncConfigurationBuilder(user)
                 .testSchema(SyncStringOnly::class.java)
                 .clientResetHandler { session: SyncSession, error: ClientResetRequiredError ->
@@ -151,12 +191,12 @@ class SessionTests {
 
     // Check that we can manually execute the Client Reset.
     @Test
+    @Deprecated("clientResetHandler deprecated in favor of syncClientResetStrategy")
     fun errorHandler_manualExecuteClientReset() = looperThread.runBlocking {
         val resources = ResourceContainer()
 
         val config = configFactory.createSyncConfigurationBuilder(user)
                 .testSchema(SyncStringOnly::class.java)
-                .clientResyncMode(ClientResyncMode.MANUAL)
                 .clientResetHandler { _: SyncSession, error: ClientResetRequiredError ->
                     try {
                         error.executeClientReset()
@@ -183,10 +223,10 @@ class SessionTests {
 
     // Check that we can use the backup SyncConfiguration to open the Realm.
     @Test
+    @Deprecated("clientResetHandler deprecated in favor of syncClientResetStrategy")
     fun errorHandler_useBackupSyncConfigurationForClientReset() = looperThread.runBlocking {
         val resources = ResourceContainer()
         val config = configFactory.createSyncConfigurationBuilder(user)
-                .clientResyncMode(ClientResyncMode.MANUAL)
                 .schema(SyncStringOnly::class.java)
                 .clientResetHandler { _: SyncSession?, error: ClientResetRequiredError ->
                     // Execute Client Reset
@@ -230,11 +270,11 @@ class SessionTests {
     // this might be the case if the user decide to act upon the client reset later (providing s/he
     // persisted the location of the file)
     @Test
+    @Deprecated("clientResetHandler deprecated in favor of syncClientResetStrategy")
     fun errorHandler_useBackupSyncConfigurationAfterClientReset() = looperThread.runBlocking {
         val resources = ResourceContainer()
         val config = configFactory.createSyncConfigurationBuilder(user)
                 .modules(SyncStringOnlyModule())
-                .clientResyncMode(ClientResyncMode.MANUAL)
                 .clientResetHandler { session: SyncSession?, error: ClientResetRequiredError ->
                     // Execute Client Reset
                     resources.close()
@@ -292,12 +332,12 @@ class SessionTests {
 
     // make sure the backup file Realm is encrypted with the same key as the original synced Realm.
     @Test
+    @Deprecated("clientResetHandler deprecated in favor of syncClientResetStrategy")
     fun errorHandler_useClientResetEncrypted() = looperThread.runBlocking {
         val resources = ResourceContainer()
 
         val randomKey = TestHelper.getRandomKey()
         val config = configFactory.createSyncConfigurationBuilder(user)
-                .clientResyncMode(ClientResyncMode.MANUAL)
                 .encryptionKey(randomKey)
                 .modules(SyncStringOnlyModule())
                 .clientResetHandler { session: SyncSession?, error: ClientResetRequiredError ->
@@ -329,6 +369,220 @@ class SessionTests {
                     looperThread.testComplete()
                 }
                 .build()
+
+        val realm = Realm.getInstance(config)
+        realm.executeTransaction {
+            realm.createObject(SyncStringOnly::class.java, ObjectId()).chars = "Foo"
+        }
+        resources.add(realm)
+
+        // Trigger error
+        user.app.sync.simulateClientReset(realm.syncSession)
+    }
+
+
+    // Check that a Client Reset is correctly reported.
+    @Test
+    fun errorHandler_manuallyRecoverClientResetReported() = looperThread.runBlocking {
+        val config = configFactory.createSyncConfigurationBuilder(user)
+            .testSchema(SyncStringOnly::class.java)
+            .syncClientResetStrategy { session: SyncSession, error: ClientResetRequiredError ->
+                val filePathFromError = error.originalFile.absolutePath
+                val filePathFromConfig = session.configuration.path
+                assertEquals(filePathFromError, filePathFromConfig)
+                assertFalse(error.backupFile.exists())
+                assertTrue(error.originalFile.exists())
+                // Note, this error message is just the one created by ObjectStore for testing
+                // The server will send a different message. This just ensures that we don't
+                // accidentially modify or remove the message.
+                assertEquals("Simulate Client Reset", error.message)
+                looperThread.testComplete()
+            }
+            .build()
+
+        val realm = Realm.getInstance(config)
+        looperThread.closeAfterTest(realm)
+
+        // Trigger error
+        user.app.sync.simulateClientReset(realm.syncSession)
+    }
+
+    // Check that we can manually execute the Client Reset.
+    @Test
+    fun errorHandler_manuallyRecoverExecuteClientReset() = looperThread.runBlocking {
+        val resources = ResourceContainer()
+
+        val config = configFactory.createSyncConfigurationBuilder(user)
+            .testSchema(SyncStringOnly::class.java)
+            .syncClientResetStrategy { _: SyncSession, error: ClientResetRequiredError ->
+                try {
+                    error.executeClientReset()
+                    fail("All Realms should be closed before executing Client Reset can be allowed")
+                } catch (ignored: IllegalStateException) {
+                }
+
+                // Execute Client Reset
+                resources.close()
+                error.executeClientReset()
+
+                // Validate that files have been moved
+                assertFalse(error.originalFile.exists())
+                assertTrue(error.backupFile.exists())
+                looperThread.testComplete()
+            }
+            .build()
+        val realm = Realm.getInstance(config)
+        resources.add(realm)
+
+        // Trigger error
+        user.app.sync.simulateClientReset(realm.syncSession)
+    }
+
+    // Check that we can use the backup SyncConfiguration to open the Realm.
+    @Test
+    fun errorHandler_manuallyUseBackupSyncConfigurationForClientReset() = looperThread.runBlocking {
+        val resources = ResourceContainer()
+        val config = configFactory.createSyncConfigurationBuilder(user)
+            .schema(SyncStringOnly::class.java)
+            .syncClientResetStrategy { _: SyncSession?, error: ClientResetRequiredError ->
+                // Execute Client Reset
+                resources.close()
+                error.executeClientReset()
+
+                // Validate that files have been moved
+                assertFalse(error.originalFile.exists())
+                assertTrue(error.backupFile.exists())
+                val backupRealmConfiguration = error.backupRealmConfiguration
+                assertNotNull(backupRealmConfiguration)
+                assertFalse(backupRealmConfiguration is SyncConfiguration)
+                assertTrue(backupRealmConfiguration.isRecoveryConfiguration)
+                Realm.getInstance(backupRealmConfiguration).use { backupRealm ->
+                    assertFalse(backupRealm.isEmpty)
+                    assertEquals(1, backupRealm.where(SyncStringOnly::class.java).count())
+                    assertEquals("Foo", backupRealm.where(SyncStringOnly::class.java).findAll().first()!!.chars)
+                }
+
+                // opening a Dynamic Realm should also work
+                DynamicRealm.getInstance(backupRealmConfiguration).use { dynamicRealm ->
+                    assertNotNull(dynamicRealm.schema.get(SyncStringOnly.CLASS_NAME))
+                    val all = dynamicRealm.where(SyncStringOnly.CLASS_NAME).findAll()
+                    assertEquals(1, all.size.toLong())
+                    assertEquals("Foo", all.first()!!.getString(SyncStringOnly.FIELD_CHARS))
+                }
+                looperThread.testComplete()
+            }
+            .build()
+        val realm = Realm.getInstance(config)
+        realm.executeTransaction {
+            realm.createObject(SyncStringOnly::class.java, ObjectId()).chars = "Foo"
+        }
+        resources.add(realm)
+
+        // Trigger error
+        user.app.sync.simulateClientReset(realm.syncSession)
+    }
+
+    // Check that we can open the backup file without using the provided SyncConfiguration,
+    // this might be the case if the user decide to act upon the client reset later (providing s/he
+    // persisted the location of the file)
+    @Test
+    fun errorHandler_manuallyUseBackupSyncConfigurationAfterClientReset() = looperThread.runBlocking {
+        val resources = ResourceContainer()
+        val config = configFactory.createSyncConfigurationBuilder(user)
+            .modules(SyncStringOnlyModule())
+            .syncClientResetStrategy { session: SyncSession?, error: ClientResetRequiredError ->
+                // Execute Client Reset
+                resources.close()
+                error.executeClientReset()
+
+                // Validate that files have been moved
+                assertFalse(error.originalFile.exists())
+                assertTrue(error.backupFile.exists())
+                val backupFile = error.backupFile.absolutePath
+
+                // this SyncConf doesn't specify any module, it will throw a migration required
+                // exception since the backup Realm contain only StringOnly table
+                var backupRealmConfiguration = SyncConfiguration.forRecovery(backupFile)
+                assertFailsWith<RealmMigrationNeededException> {
+                    Realm.getInstance(backupRealmConfiguration)
+                }
+
+                // opening a DynamicRealm will work though
+                DynamicRealm.getInstance(backupRealmConfiguration).use { dynamicRealm ->
+                    assertNotNull(dynamicRealm.schema.get(SyncStringOnly.CLASS_NAME))
+                    val all = dynamicRealm.where(SyncStringOnly.CLASS_NAME).findAll()
+                    assertEquals(1, all.size.toLong())
+                    assertEquals("Foo", all.first()!!.getString(SyncStringOnly.FIELD_CHARS))
+                    // make sure we can't write to it (read-only Realm)
+                    assertFailsWith<java.lang.IllegalStateException> {
+                        dynamicRealm.beginTransaction()
+                    }
+                }
+
+                assertFailsWith<IllegalArgumentException> {
+                    SyncConfiguration.forRecovery(backupFile, null, SyncStringOnly::class.java)
+                }
+
+                // specifying the module will allow to open the typed Realm
+                backupRealmConfiguration = SyncConfiguration.forRecovery(backupFile, null, SyncStringOnlyModule())
+                Realm.getInstance(backupRealmConfiguration).use { backupRealm ->
+                    assertFalse(backupRealm.isEmpty)
+                    assertEquals(1, backupRealm.where(SyncStringOnly::class.java).count())
+                    val allSorted = backupRealm.where(SyncStringOnly::class.java).findAll()
+                    assertEquals("Foo", allSorted[0]!!.chars)
+                }
+                looperThread.testComplete()
+            }
+            .build()
+
+        val realm = Realm.getInstance(config)
+        realm.executeTransaction {
+            it.createObject(SyncStringOnly::class.java, ObjectId()).chars = "Foo"
+        }
+        resources.add(realm)
+
+        // Trigger error
+        user.app.sync.simulateClientReset(realm.syncSession)
+    }
+
+    // make sure the backup file Realm is encrypted with the same key as the original synced Realm.
+    @Test
+    fun errorHandler_manuallyUseClientResetEncrypted() = looperThread.runBlocking {
+        val resources = ResourceContainer()
+
+        val randomKey = TestHelper.getRandomKey()
+        val config = configFactory.createSyncConfigurationBuilder(user)
+            .encryptionKey(randomKey)
+            .modules(SyncStringOnlyModule())
+            .syncClientResetStrategy { session: SyncSession?, error: ClientResetRequiredError ->
+                // Execute Client Reset
+                resources.close()
+                error.executeClientReset()
+                var backupRealmConfiguration = error.backupRealmConfiguration
+
+                // can open encrypted backup Realm
+                Realm.getInstance(backupRealmConfiguration).use { backupEncryptedRealm ->
+                    assertEquals(1, backupEncryptedRealm.where(SyncStringOnly::class.java).count())
+                    val allSorted = backupEncryptedRealm.where(SyncStringOnly::class.java).findAll()
+                    assertEquals("Foo", allSorted[0]!!.chars)
+                }
+                val backupFile = error.backupFile.absolutePath
+
+                // build a conf to open a DynamicRealm
+                backupRealmConfiguration = SyncConfiguration.forRecovery(backupFile, randomKey, SyncStringOnlyModule())
+                Realm.getInstance(backupRealmConfiguration).use { backupEncryptedRealm ->
+                    assertEquals(1, backupEncryptedRealm.where(SyncStringOnly::class.java).count())
+                    val allSorted = backupEncryptedRealm.where(SyncStringOnly::class.java).findAll()
+                    assertEquals("Foo", allSorted[0]!!.chars)
+                }
+
+                // using wrong key throw
+                assertFailsWith<RealmFileException> {
+                    Realm.getInstance(SyncConfiguration.forRecovery(backupFile, TestHelper.getRandomKey(), SyncStringOnlyModule()))
+                }
+                looperThread.testComplete()
+            }
+            .build()
 
         val realm = Realm.getInstance(config)
         realm.executeTransaction {